+++ src/jvmMain/kotlin/creals/impl/IntegerConstantReal.kt
@@ -1,14 +1,13 @@
 package garden.ephemeral.calculator.creals.impl
 
+import garden.ephemeral.calculator.bigint.BigInt
 import garden.ephemeral.calculator.creals.Real
-import garden.ephemeral.calculator.creals.util.scale
-import java.math.BigInteger
 
 /**
  * Representation of an integer constant.
  */
-internal class IntegerConstantReal(val value: BigInteger) : Real() {
-    override fun approximate(precision: Int): BigInteger {
+internal class IntegerConstantReal(val value: BigInt) : Real() {
+    override fun approximate(precision: Int): BigInt {
         return value.scale(-precision)
     }
 }
